# dist/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Beldex Exchange",
        "description": "This is a core api server endpoints.  You can find out more about Beldex at [https://beldex.io](https://beldex.io).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "sathish.kumar@beldex.org"
        }
    },
    "host": "coreapialpha.beldex.io",
    "basePath": "/api/v1/",
...
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "user/activation/{hash}": {
            "get": {
                "tags": [
                    "users"
                ],
                "description": "Returns response based on hash key",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Account activation",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/activation"
                        }
                    }
                ]
            }
        },
        "user/login": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "Returns login credentials correct or not",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "ID",
                        "in": "body",
                        "description": "Account login",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/login"
                        }
                    }
                ]
            }
        },
        "user/validate/otp": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "ValidateOtp",
                "operationId": "find",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "settings",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/validate"
                        }
                    }
                ]
            }
        },
        "user/resend/otp": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "resendOtp",
                "operationId": "findOneAndUpdate",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "settings",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/resend"
                        }
                    }
                    
                ]
            }
        },
        "user/forget-password": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "send link to your registered email address",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "ID",
                        "in": "body",
                        "description": "Forget password",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/forgotpassword"
                        }
                    }
                ]
            }
        },
        "user/reset-password/{hash}": {
            "get": {
                "tags": [
                    "users"
                ],
                "description": "Check correct hash",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Reset password",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/resetpasswordhash"
                        }
                    }
                ]
            }
        },
        "user/reset-password": {
            "patch": {
                "tags": [
                    "users"
                ],
                "description": "Reset password",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "ID",
                        "in": "body",
                        "description": "Reset password",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/resetpassword"
                        }
                    }
                ]
            }
        },
        "user/change-password": {
            "patch": {
                "tags": [
                    "users"
                ],
                "description": "Change password",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "JSON data",
                        "in": "body",
                        "description": "Change password",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/changepassword"
                        }
                    }
                ]
            }
        },
        "user/whitelist-ip/{hash}": {
            "patch": {
                "tags": [
                    "users"
                ],
                "description": "Working IP check",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "ID to fetch",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/whitelist"
                        }
                    }
                ]
            }
        },
        "user/get-user-id": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "Get user id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Get user id",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/getuserid"
                        }
                    }
                ]
            }
        },
        "user/device-history": {
            "get": {
                "tags": [
                    "users"
                ],
                "description": "Get device history",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Get device history",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/devicehistory"
                        }
                    }
                ]
            }
        },
        "user/login-history": {
            "get": {
                "tags": [
                    "users"
                ],
                "description": "Get login history",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Get login history",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/loginhistory"
                        }
                    }
                ]
            }
        },
...
        "user/resend-email": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "Resend Mail",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Resend mail",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/resendmail"
                        }
                    }
                ]
            }
        },
        "user/g2f-verify": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "g2f verify",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "g2f verify",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Assets",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/g2f-verify"
                        }
                    }
                ]
            }
        },
        "user/g2f-settings": {
            "patch": {
                "tags": [
                    "users"
                ],
                "description": "Get g2f settings",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "g2f settings",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "g2f settings",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/g2f-settings"
                        }
                    }
                ]
            }
        },
        "user/settings": {
            "patch": {
                "tags": [
                    "users"
                ],
                "description": "Get user settings",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Setting",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/settings"
                        }
                    }
                ]
            }
        },
        "user/disable": {
            "patch": {
                "tags": [
                    "users"
                ],
                "description": "Get user settings",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "Setting",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/disable"
                        }
                    }
                ]
            }
        },
        "user/token": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": " Create Refresh Token",
                "operationId": "",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Setting",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/refreshToken"
                        }
                    }
                ]
            }
        },
        "user/logout": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "User Logout and add logout time",
                "operationId": "findOneAndUpdate",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Setting",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/logout"
                        }
                    }
                ]
            }
        },
        "user/whitelist": {
            "delete": {
                "tags": [
                    "users"
                ],
                "description": "Delete whitelist user in Db",
                "operationId": "updateMany",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "settings",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "settings",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/deleteWhitList"
                        }
                    }
                ]
            }
        },
        "user/market": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "Add market in Db",
                "operationId": "Create",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "settings",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/market"
                        }
                    }
                ]
            }
        },
...
        "user/favourite": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "Add market in favourite list",
                "operationId": "Create",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "settings",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/favourite"
                        }
                    },
                    {

                        "name": "Authorization",
                        "in": "header",
                        "description": "settings",
                        "required": true,
                        "type": "string"
                    }
                ]
            },
            "patch": {
                "tags": [
                    "users"
                ],
                "description": "remove market in favourite list",
                "operationId": "patch",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Parameters",
                        "in": "body",
                        "description": "settings",
                        "required": true,
                        "type": "string",
                        "schema": {
                            "$ref": "#/definitions/favourite"
                        }
                    },
                    {

                        "name": "Authorization",
                        "in": "header",
                        "description": "settings",
                        "required": true,
                        "type": "string"
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "login": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "email": "saradaparamasivan@gmail.com",
                            "password": "1234567S",
                            "is_browser": true,
                            "is_mobile": false,
                            "os": "Linux",
                            "os_byte": "x86_64",
                            "browser": "Chrome",
                            "browser_version": "71.0.3578.11",
                            "ip": "171.78.139.181",
                            "city": "Chennai",
                            "region": "Tamil Nadu",
                            "country": "India"
                        }
                    }
                }
            }
        },
        "validate": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "otp":"BEL-12345",
                            "device_id":"123456789",
                            "user_id":"123456789",
                            "region":"Tamil Nadu",
                            "city":"Chennai",
                            "ip":"171.78.139.181"

                        }
                    }
                }
            }
        },
        "resend": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "user_id":"123456789"
                        }
                    }
                }
            }
        },
        "forgotpassword": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "attributes": {
                            "email": "xyz@gmail.com"
                        }
                    }
                }
            }
        },
        "resetpassword": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "id": "5c90dd8e6410cd0378d8fd7e",
                        "attributes": {
                            "password": "tesR123456",
                            "password_confirmation": "tesR123456"
                        }
                    }
                }
            }
        },
        "changepassword": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "id": "5c90dd8e6410cd0378d8fd7e",
                        "attributes": {
                            "old_password":"1234567S",
                            "password": "tesR123456",
                            "password_confirmation": "tesR123456"
                        }
                    }
                }
            }
        },
        "activation": {
            "properties": {
                "code": {
                    "example": 200
                },
                "errors": {
                    "required": true,
                    "type": "string",
                    "example": "false"
                },
                "data": {}
            }
        },
        "resetpasswordhash": {
            "properties": {
                "code": {
                    "example": 202
                },
                "errors": {
                    "required": true,
                    "type": "string",
                    "example": "false"
                },
                "data": {
                    "example": {
                        "id": "your id",
                        "type": "users",
                        "attributes": {
                            "message": "Your password updated successfully."
                        }
                    }
                }
            }
        },
        "getuserid": {
            "properties": {
                "code": {
                    "example": 0
                },
                "message": {
                    "example": "Authorization successfully."
                },
                "data": {
                    "example": {
                        "user_id": 1
                    }
                }
            }
        },
        "whitelist": {
            "properties": {
                "code": {
                    "example": 202
                },
                "errors": {
                    "example": "false"
                },
                "data": {
                    "example": {
                        "id": "null",
                        "type": "users",
                        "attributes": {
                            "message": "Your IP address whitelisted Now you can able to login.."
                        }
                    }
                }
            }
        },
...
        "devicehistory": {
            "properties": {
                "code": {
                    "example": 200
                },
                "errors": {
                    "example": "false"
                },
                "data": {
                    "example": {
                        "id": "5c908652f83b10002f0f7478",
                        "type": "devices",
                        "attributes": {
                            "data": []
                        }
                    }
                }
            }
        },
        "loginhistory": {
            "properties": {
                "code": {
                    "example": 200
                },
                "errors": {
                    "example": "false"
                },
                "data": {
                    "example": {
                        "id": "5c908652f83b10002f0f7478",
                        "type": "loginHistory",
                        "attributes": {
                            "data": []
                        }
                    }
                }
            }
        },
...
        "refreshToken": {
            "properties": {
                "code": {
                    "example": 200
                },
                "errors": {
                    "example": "false"
                }
            }
        },
...
        "resendmail": {
            "properties": {
                "lang": {
                    "required": true,
                    "type": "string",
                    "example": "en"
                },
                "data": {
                    "example": {
                        "id": "5c960beebd528c10aae89c15",
                        "attributes": {
                            "type": "registration"
                        }
                    }
                }
            }
        },
...
    }
}